Co-Authored By: Jeannie Evans @jmevans0211 & Edwin del Bosque @edwindelbosque

+++ main.js
@@ -1,47 +1,34 @@
-// var btnDelete = document.querySelector('#btn-delete');
-// var btnDownvote = document.querySelector('#btn-downvote');
-// var btnGenius = document.querySelector('#btn-genius');
-// var btnPlausible = document.querySelector('#btn-plausible');
-// var btnQuality = document.querySelector('#btn-quality');
 var btnSave = document.querySelector('#btn-save');
-// var btnStar = document.querySelector('#btn-star');
-// var btnStarred = document.querySelector('#btn-star');
-// var btnSwill = document.querySelector('#btn-swill');
-// var btnUpvote = document.querySelector('#btn-upvote');
 var inputBody = document.querySelector('#input-body');
-// var inputQuality = document.querySelector('#input-quality');
-// var inputSearch = document.querySelector('#input-search')
 var inputTitle = document.querySelector('#input-title');
 var cardArea = document.querySelector('.section-bottom');
 var ideasArray = []
-// for (i = 0; i > ideasArray.length; i++) {
+var qualitiesArray = ["Swill", "Plausible", "Genius"];
+var btnMenu = document.querySelector('.icons-backdrop');
 
-//   (ideaCard + i) = document.querySelector(`idea-card-${i}`);
-//   (ideaCard + i).addEventListener('blur', updateCard);
-// }
-var id = Date.now();
-
-// btnGenius.addEventListener('click', );
-// btnPlausible.addEventListener('click', );
-// btnQuality.addEventListener('click', toggleArrow);
 btnSave.addEventListener('click', addIdea);
-cardArea.addEventListener('click', toggleStar);
-// btnSwill.addEventListener('click', );
-
 cardArea.addEventListener('focusout', handleFocusOut);
 cardArea.addEventListener('keydown', handleTextEdit);
 cardArea.addEventListener('click', handleCardButtons);
-inputTitle.addEventListener('keyup', handleSaveBtn);
-
+btnMenu.addEventListener('click', toggleMenu)
 inputBody.addEventListener('keyup', handleSaveBtn);
+inputTitle.addEventListener('keyup', handleSaveBtn);
 window.addEventListener('DOMContentLoaded', repopulateIdeasArray);
 
-// handleBottom() {
-// 	if (e.)
-// 		make if functions to target every button/image
-// }
+
 function handleCardButtons(e) {
-  deleteCard(e);
+  if (e.target.id === 'btn-upvote') {
+    upvoteQuality(e);
+  }
+  if (e.target.id === 'btn-downvote') {
+    downvoteQuality(e);
+  }
+  if (e.target.id === 'btn-delete') {
+    deleteCard(e);
+  }
+  if (e.target.id === 'btn-star') {
+    toggleStar(e);
+  }
 }
 
 function addIdea(e) {
@@ -50,20 +37,18 @@ function addIdea(e) {
 	var idea = new Idea(Date.now(), inputTitle.value, inputBody.value, false, 0);
 	ideasArray.push(idea);
 	idea.saveToStorage(ideasArray);
-	handleSaveBtn();
 	addCard(idea);
 	inputTitle.value = "";
 	inputBody.value = "";
+	handleSaveBtn();
 }
 
-handleSaveBtn();
-
 function addCard(object) {
+  var changeStar = object.star ? 'images/star-active.svg' : 'images/star.svg';
   var numOfIdeas = ideasArray.length;
 	cardArea.insertAdjacentHTML('afterbegin', `<article class="idea-card" data-id="${object.id}">
         <header>
-          <img class="img-star star" src="images/star.svg" alt="star" id="btn-star">
-          <img class="img-star star-active hidden" src="images/star-active.svg" alt="active star" id="btn-star-active">
+          <img class="img-star star" src="${changeStar}" alt="star" id="btn-star">
           <img class="img-delete" src="images/delete.svg" alt="delete" id="btn-delete">
         </header>
         <div class="card-content">
@@ -72,7 +57,7 @@ function addCard(object) {
         </div>
         <footer>
           <img class="img-upvote" src="images/upvote.svg" alt="upvote" id="btn-upvote">
-          <p class="quality-text">Quality: ${object.quality}</p>
+          <p class="quality-text">Quality: ${qualitiesArray[object.quality]}</p>
           <img class="img-downvote" src="images/downvote.svg" alt="downvote" id="btn-downvote">    
         </footer>
       </article>`);
@@ -90,19 +75,48 @@ function repopulateIdeasArray() {
   }
 }
 
+function upvoteQuality(e) {
+  e.target.closest('.idea-card');
+  var idea = findIdea(e);
+
+  if (idea.quality < qualitiesArray.length - 1) {
+    var newQuality = idea.quality + 1;
+    idea.updateQuality(newQuality);
+    qualityDisplay(e);
+    idea.saveToStorage(ideasArray);
+  }
+}
+
+function downvoteQuality(e) {
+  e.target.closest('.idea-card');
+  var idea = findIdea(e);
+
+  if (idea.quality > 0) {
+    var newQuality = idea.quality - 1;
+    idea.updateQuality(newQuality);
+    qualityDisplay(e)
+    idea.saveToStorage(ideasArray);
+  }
+}
+
+function qualityDisplay(e) {
+  var qualityDisplay = e.target.closest('.idea-card').querySelector('.quality-text');
+  var idea = findIdea(e);
+  var ideaQuality = qualitiesArray[idea.quality];
+
+  qualityDisplay.innerText = `Quality: ${ideaQuality}`
+}
+
 function deleteCard(e) {
-  console.log('inside delete');
-  if (e.target.id === 'btn-delete'){
     e.target.closest('.idea-card').remove(); 
     var idea = findIdea(e);
     idea.deleteFromStorage(ideasArray);
-  }
 }
 
 function findIdea(e) {
   var ideaId = e.target.closest('.idea-card').getAttribute('data-id');
   var idea = ideasArray.find(function(idea) {
-   return parseInt(idea.id) === parseInt(ideaId);
+   return idea.id === parseInt(ideaId);
   });
   return idea;
 }
@@ -119,10 +133,12 @@ function handleFocusOut(e) {
 }
 
 function toggleStar(e) {
-  if (e.target.id === 'btn-star') {
-    var on = document.querySelector('.star-active').classList.remove('hidden');
-    var off = document.querySelector('.star').classList.add('hidden');
-  }
+    var idea = findIdea(e);
+    idea.updateStar();
+    var changeStar = idea.star ? 'images/star-active.svg' : 'images/star.svg';
+    e.target.setAttribute('src', changeStar);
+
+    idea.saveToStorage(ideasArray);
 }
 
 function handleTextEdit(e) {
@@ -139,15 +155,24 @@ function handleTextEdit(e) {
  } 
 
 function handleSaveBtn() {
-  console.log("inputTitle", inputTitle.value);
-  console.log("inputBody", inputBody.value);
 	btnSave.disabled = !inputTitle.value || !inputBody.value;
 }
-  // console.log(title);
-  //return enter key saves changes
-  //assign the new fields to the property values on DOM
-  //get the objects with the changes
-  //update the array to include new changes
-  //pass the array to local storage udpateIdea() to update data model
 
+function toggleMenu(e) {
+  var btnMenu = document.querySelector('.icons-backdrop');
+  var classes = ["show-aside", "hide-aside"]
+  var iconOne = ["burger-icon-1", "x-icon-1"]
+  var iconTwo = ["burger-icon-1", "x-icon-2"]
+  var iconThree = ["burger-icon-1", "x-icon-3"]
 
+  function animationLoop(classChange, element) {
+    for (var i = 0; i < classChange.length; i++){
+    document.querySelector(element).classList.toggle(classChange[i])
+    }
+  }
+
+  animationLoop(classes, 'aside');
+  animationLoop(iconOne, '#burger-icon-1');
+  animationLoop(iconTwo, '#burger-icon-2');
+  animationLoop(iconThree, '#burger-icon-3');
+  }
